+++ UI/src/lib/models/product.ts
