+++ UI/src/lib/models/productType.ts
